Added specific editor application and rename some files and classes
