+++ projects/AngerManagement/src/EditorApplication.cpp
@@ -1,0 +1,14 @@
+﻿#include "EditorApplication.h"
+
+namespace AM
+{
+	EditorApplication::EditorApplication(int& argc, char** argv) : QApplication(argc, argv)
+	{
+	
+	}
+
+	EditorApplication::~EditorApplication()
+	{
+	
+	}
+}
