+++ projects/AngerManagement/src/EditorApplication.h
@@ -1,0 +1,18 @@
+﻿#pragma once
+
+#include <QApplication>
+
+namespace AM
+{
+	class EditorApplication : public QApplication
+	{
+		Q_OBJECT
+
+	public:
+		EditorApplication(int& argc, char** argv);
+		~EditorApplication();
+
+	private:
+		
+	};
+}
