+++ projects/AngerManagement/src/EditorWindow.qrc
